pyproject.toml
[project]
name = "sheets-search-mcp"
version = "0.1.0"
description = "MCP server and Slack bot for querying Google Sheets data with Claude"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.9.0",
    "gspread>=6.0.0",
    "google-api-python-client>=2.0.0",
    "slack-bolt>=1.18.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
